LPS-105380 Service Buidler *abc-service should not use abc.service.internal, but just abc.internal
